Refactor mdsal-binding-generator artifacts
Generator artifacts have a layout which belies their actual purpose,
as exposed after mdsal-binding-generator-impl refactor.

mdsal-binding-generator-impl is renamed to mdsal-binding-generator,
absorbing BindingGenerator from the API artifact.

mdsal-binding-generator-{api,util} are renamed to
mdsal-binding-model-{api,ri}, with generator-specific utilities being
moved to mdsal-binding-generator.

This change makes it easier to integrate
mdsal-binding-java-api-generator, as we can directly depend on
mdsal-binding-generator.

+++ binding/mdsal-binding-java-api-generator/src/main/java/org/opendaylight/mdsal/binding/java/api/generator/LengthGenerator.java
@@ -14,7 +14,7 @@ import java.util.Set;
 import org.eclipse.jdt.annotation.NonNull;
 import org.eclipse.jdt.annotation.Nullable;
 import org.opendaylight.mdsal.binding.model.api.Type;
-import org.opendaylight.mdsal.binding.model.util.Types;
+import org.opendaylight.mdsal.binding.model.ri.Types;
 import org.opendaylight.yangtools.yang.model.api.type.LengthConstraint;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -36,7 +36,7 @@ final class LengthGenerator {
 
         for (Range<Integer> l : constraints) {
             // We have to deal with restrictions being out of integer's range
-            final String expr = createExpression(l.lowerEndpoint().intValue(), l.upperEndpoint().intValue());
+            final String expr = createExpression(l.lowerEndpoint(), l.upperEndpoint());
             if (expr == null) {
                 // This range is implicitly capped by String/byte[] length returns
                 LOG.debug("Constraint {} implied by int type value domain, skipping", l);
